Properly quote variables
When I wrote this script I was unaware that "case" differentiated
between quoted and unquoted variables in case labels, and I was unaware
that "local" takes an argument list like any other command (rather than
a list of variable declarations). I was also unaware of "cd --" in dash.
I've fixed relevant portions of the script to ensure that it behaves
correctly.