Do not search for a VCS if PWD is not absolute
If PWD doesn't begin with "/", the VCS search loop will continue
indefinitely. It now bails out early instead of hanging in this case.
I doubt PWD will ever be a non-absolute path, but I like to program
defensively.

CVS will still be detected regardless of the value of PWD, since it adds
its metadata to every tracked directory rather than just the root of the
repository.